+++ src/test/java/eu/dissco/core/handlemanager/service/DoiServiceTest.java
@@ -1,6 +1,9 @@
 package eu.dissco.core.handlemanager.service;
 
 import static eu.dissco.core.handlemanager.domain.FdoProfile.PRIMARY_SPECIMEN_OBJECT_ID;
+import static eu.dissco.core.handlemanager.testUtils.TestUtils.CREATED;
+import static eu.dissco.core.handlemanager.testUtils.TestUtils.HANDLE;
+import static eu.dissco.core.handlemanager.testUtils.TestUtils.HANDLE_ALT;
 import static eu.dissco.core.handlemanager.testUtils.TestUtils.HANDLE_DOMAIN;
 import static eu.dissco.core.handlemanager.testUtils.TestUtils.MAPPER;
 import static eu.dissco.core.handlemanager.testUtils.TestUtils.RECORD_TYPE_DS;
@@ -15,24 +18,74 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mockStatic;
 
 import eu.dissco.core.handlemanager.Profiles;
 import eu.dissco.core.handlemanager.domain.repsitoryobjects.HandleAttribute;
 import eu.dissco.core.handlemanager.domain.requests.vocabulary.ObjectType;
 import eu.dissco.core.handlemanager.exceptions.InvalidRequestException;
+import eu.dissco.core.handlemanager.properties.ProfileProperties;
+import eu.dissco.core.handlemanager.repository.PidRepository;
+import java.nio.charset.StandardCharsets;
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.List;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.MockedStatic;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.context.ActiveProfiles;
 
+@ExtendWith(MockitoExtension.class)
 @ActiveProfiles(profiles = Profiles.DOI)
-class DoiServiceTest extends PidServiceTest {
+class DoiServiceTest {
+
+  @Mock
+  private PidRepository pidRepository;
+  @Mock
+  private FdoRecordService fdoRecordService;
+  @Mock
+  private PidNameGeneratorService pidNameGeneratorService;
+  @Mock
+  private ProfileProperties profileProperties;
+  private PidService service;
+  private List<byte[]> handles;
+  private MockedStatic<Instant> mockedStatic;
+  private MockedStatic<Clock> mockedClock;
 
   @BeforeEach
-  void initService() {
+  void setup() {
+    initTime();
+    initHandleList();
     service = new DoiService(pidRepository, fdoRecordService, pidNameGeneratorService, MAPPER,
         profileProperties);
+  }
+
+  private void initTime() {
+    Clock clock = Clock.fixed(CREATED, ZoneOffset.UTC);
+    Instant instant = Instant.now(clock);
+    mockedStatic = mockStatic(Instant.class);
+    mockedStatic.when(Instant::now).thenReturn(instant);
+    mockedStatic.when(() -> Instant.from(any())).thenReturn(instant);
+    mockedClock = mockStatic(Clock.class);
+    mockedClock.when(Clock::systemUTC).thenReturn(clock);
+  }
+
+  private void initHandleList() {
+    handles = new ArrayList<>();
+    handles.add(HANDLE.getBytes(StandardCharsets.UTF_8));
+    handles.add(HANDLE_ALT.getBytes(StandardCharsets.UTF_8));
+  }
+
+  @AfterEach
+  void destroy() {
+    mockedStatic.close();
+    mockedClock.close();
   }
 
   @Test
