+++ src/test/java/eu/dissco/core/handlemanager/service/HandleServiceTest.java
@@ -65,6 +65,7 @@ import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
+import static org.mockito.Mockito.mockStatic;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import eu.dissco.core.handlemanager.Profiles;
@@ -77,25 +78,71 @@ import eu.dissco.core.handlemanager.domain.requests.vocabulary.PrimaryObjectIdTy
 import eu.dissco.core.handlemanager.exceptions.InvalidRequestException;
 import eu.dissco.core.handlemanager.exceptions.PidCreationException;
 import eu.dissco.core.handlemanager.exceptions.PidResolutionException;
+import eu.dissco.core.handlemanager.properties.ProfileProperties;
+import eu.dissco.core.handlemanager.repository.PidRepository;
 import java.nio.charset.StandardCharsets;
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.context.ActiveProfiles;
 
 @ExtendWith(MockitoExtension.class)
 @ActiveProfiles(profiles = Profiles.HANDLE)
-class HandleServiceTest extends PidServiceTest {
+class HandleServiceTest {
+
+  @Mock
+  private PidRepository pidRepository;
+  @Mock
+  private FdoRecordService fdoRecordService;
+  @Mock
+  private PidNameGeneratorService pidNameGeneratorService;
+  @Mock
+  private ProfileProperties profileProperties;
+  private PidService service;
+  private List<byte[]> handles;
+  private MockedStatic<Instant> mockedStatic;
+  private MockedStatic<Clock> mockedClock;
 
   @BeforeEach
-  void initService() {
+  void setup() {
+    initTime();
+    initHandleList();
     service = new HandleService(pidRepository, fdoRecordService, pidNameGeneratorService, MAPPER,
         profileProperties);
   }
+
+  private void initTime() {
+    Clock clock = Clock.fixed(CREATED, ZoneOffset.UTC);
+    Instant instant = Instant.now(clock);
+    mockedStatic = mockStatic(Instant.class);
+    mockedStatic.when(Instant::now).thenReturn(instant);
+    mockedStatic.when(() -> Instant.from(any())).thenReturn(instant);
+    mockedClock = mockStatic(Clock.class);
+    mockedClock.when(Clock::systemUTC).thenReturn(clock);
+  }
+
+  private void initHandleList() {
+    handles = new ArrayList<>();
+    handles.add(HANDLE.getBytes(StandardCharsets.UTF_8));
+    handles.add(HANDLE_ALT.getBytes(StandardCharsets.UTF_8));
+  }
+
+  @AfterEach
+  void destroy() {
+    mockedStatic.close();
+    mockedClock.close();
+  }
+
 
   @Test
   void testResolveSingleRecord() throws Exception {
